+++ CHANGELOG.md
@@ -25,6 +25,7 @@
 - Allow CotEditor Syntax files and multiple replacement definitions to be imported by dropping them onto their respective lists.
 - Add a context menu to the list in the Mode settings pane.
 - Display more accurate error messages in dialogs.
+- Display the number of definition errors as a badge in the sidebar of the syntax editor.
 - [dev] Tweak the accent color.
 - [dev] Update the build environment to Xcode 27 Beta 2 (Swift 6.4, macOS 27 SDK).